refactor: interfaces -> ui 패키지명 수정
